@@ -5,10 +5,9 @@ import os
 import telebot
 
 from db import Database, Password, InsertQuery
-from db.query import SimpleSelectQuery
+from db.query import DeleteQuery, JoinSelectQuery, SimpleSelectQuery, Where
 
 import utils
-import exceptions
 
 
 TOKEN = os.getenv("TG_TOKEN")
@@ -32,53 +31,27 @@ def send_welcome(message: telebot.types.Message):
     )
 
 
-# @bot.message_handler(commands=["add"])
-# def add_expense(message):
-#     try:
-#         tokens = message.text.split()
-#         if len(tokens) < 2:
-#             raise exceptions.IncorrectArgumentCount
-#         payload = {"cat_id": 1, "cat_description": ""}
-#         for index, token in enumerate(tokens):
-#             if index == 0:
-#                 continue
-#             elif index == 1:
-#                 payload["amount"] = float(token)
-#             elif index == 2:
-#                 known_categories = db.fetchall("expense_categories")
-#                 for known_category in known_categories:
-#                     if token.upper() == known_category[1].upper():
-#                         payload["cat_id"] = known_category[0]
-#                         break
-#             else:
-#                 payload["cat_description"] = " ".join(tokens[index:])
-#                 break
-#     except Exception:
-#         bot.send_message(
-#             message.chat.id,
-#             "Incorrect parameters specified!\nUsage: /add <amount> <categorie> <description>",
-#         )
-#     else:
-#         db.insert("expenses", payload)
-#         bot.send_message(message.chat.id, f"Added {payload['amount']} EUR")
-
 @bot.message_handler(commands=["add"])
 def add_expense(message: telebot.types.Message):
+    """Handles /add command. Adds new expenses to the database."""
     tokens = message.text.split()
 
-    for record in db.execute(SimpleSelectQuery("expense_categories")):
-        cat_id = 1
-        if tokens[2] in record[1]:
-            cat_id = record[0]
-            break
+    match tokens:
+        case ["/add", amount]:
+            query = InsertQuery("expenses", {"amount": amount})
+        case ["/add", amount, category]:
+            query = InsertQuery(
+                "expenses", {"amount": amount, "cat_id": utils.get_category_id(db, category)})
+        case ["/add", amount, category, *description]:
+            query = InsertQuery("expenses", {"amount": amount, "cat_id": utils.get_category_id(
+                db, category), "cat_description": " ".join(description)})
+        case _:
+            bot.send_message(
+                message.chat.id, "Incorrect parameters specified!\nUsage: /add <amount> <categorie> <description>")
+            return
 
-    query = InsertQuery("expenses", {"amount": tokens[1], "cat_id": cat_id, "cat_description": " ".join(tokens[3:])})
     db.execute(query)
-
-
-@bot.message_handler(commands=["add-cat"])
-def add_expense_category(message: telebot.types.Message):
-    ...
+    bot.send_message(message.chat.id, f"Added {amount} EUR")
 
 
 @bot.message_handler(commands=["top"])
@@ -94,25 +67,56 @@ def show_latest_expenses(message: telebot.types.Message):
         limit = 5
     finally:
         query = db.get_latest("expenses", limit)
-        table = utils.generate_table(headers=("Amount", "Category"), content=query)
+        table = utils.generate_table(
+            headers=("Amount", "Category"), content=query)
         bot.send_message(message.chat.id, table)
 
 
 @bot.message_handler(commands=["day", "month", "year"])
 def show_interval_expenses(message: telebot.types.Message):
+    """
+    Handles /day, /month and /year commands.
+    Responsible for generating daily, monthly or yearly expense report.
+    """
     interval = message.text.lstrip("/")
-    query = db.get_inverval(
-        "expenses as t", interval, "t.creation_timestamp::date, t.amount, jt.cat_name"
-    )
-    table = utils.generate_table(headers=("Date", "Amount", "Category"), content=query)
+    query = JoinSelectQuery(
+        tables=({"name": "expenses", "alias": "e"}, {"name": "expense_categories", "alias": "ec"}),
+        columns=("e.creation_timestamp::date", "e.amount", "ec.cat_name"),
+        equation="e.cat_id=ec.id",
+        where=Where(f"creation_timestamp > now() - '1 {interval}'::interval"))
+    data = db.execute(query)
+    table = utils.generate_table(headers=("Date", "Amount (EUR)", "Category"), content=data)
     bot.send_message(message.chat.id, table)
 
 
 @bot.message_handler(commands=["cat"])
-def show_categories(message: telebot.types.Message):
-    query = db.fetchall("expense_categories", "cat_name")
-    table = utils.generate_table(headers=("Categories",), content=query)
-    bot.send_message(message.chat.id, table)
+def categories(message: telebot.types.Message):
+    """Handles /cat command. Shows all known expense categories."""
+    tokens = message.text.split()
+
+    match tokens:
+        case ["/cat", "add", name]:
+            db.execute(InsertQuery("expense_categories", {"cat_name": name}))
+            bot.send_message(message.chat.id, f"Category {name} added")
+        case ["/cat", "del", name]:
+            db.execute(DeleteQuery("expense_categories",
+                       Where(f"cat_name='{name}'")))
+            bot.send_message(message.chat.id, f"Category {name} deleted")
+        case ["/cat", "list"] | ["/cat"]:
+            cats = db.execute(SimpleSelectQuery(
+                "expense_categories", ("cat_name",), where=Where("id != 1")))
+            table = utils.generate_table(headers=("Categories",), content=cats)
+            bot.send_message(message.chat.id, table)
+        case ["/cat", "help"] | _:
+            bot.send_message(message.chat.id, """
+`/cat [[sub-command]] [[arguments]]
+
+Accepted arguments:
+/cat help       - Show command help
+/cat list       - Show known expense category list
+/cat add <name> - Add new expense category
+/cat del <name> - Delete expense category`
+            """)
 
 
 bot.polling()
